index.md: Remove duplicate "TF LAVA Instance" section
This section appeared twice, closer to the beginning vs closer to the
end of doc. Location at the beginning of doc is more logical - we
should describe the official production instance before decsribing
hot to add new boards to it or setup a local development copy.
However, 2nd copy appeared to have some more up-to-date changes, so
these changes were applied to 1st copy to not be lost.

More info at the TFC-187 ticket.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: Ia666b47296aa14e362a15bdb37181e9ac5843142